sched: Fix possible overflow in cpu_cycles_to_freq()
Truncating period to u32 could lead to incorrect results. Make it u64
instead.

+++ kernel/sched/sched.h
@@ -1064,7 +1064,7 @@ static inline int cpu_max_power_cost(int cpu)
 	return cpu_rq(cpu)->cluster->max_power_cost;
 }
 
-static inline u32 cpu_cycles_to_freq(u64 cycles, u32 period)
+static inline u32 cpu_cycles_to_freq(u64 cycles, u64 period)
 {
 	return div64_u64(cycles, period);
 }
